+++ ng/practices-ui-ktbe/src/lib/util/hide-if-empty-text.stories.ts
@@ -1,6 +1,7 @@
 import { FormsModule } from '@angular/forms';
 import { Meta, moduleMetadata, StoryObj } from '@storybook/angular';
 import { PuibeHideIfEmptyTextDirective } from './hide-if-empty-text.directive';
+import { ComponentWithEntraceAnimation } from './hide-if-empty-text.stories.with-entrance-animation';
 
 type Args = {};
 
@@ -9,6 +10,7 @@ const meta: Meta<Args> = {
     decorators: [
         moduleMetadata({
             imports: [PuibeHideIfEmptyTextDirective, FormsModule],
+            declarations: [ComponentWithEntraceAnimation],
         }),
     ],
     render: (args) => {
@@ -21,7 +23,7 @@ const meta: Meta<Args> = {
                 <textarea [(ngModel)]="content" placeholder="Type here..." ></textarea>
 
                 <br/>
-                <p>Content goes here</p>
+                <p>Content goes here (always visible)</p>
                 <hr/>
 
                 <pre class="bg-red text-white p-10" >{{content}}</pre>
@@ -40,4 +42,50 @@ type Story = StoryObj<Args>;
 
 export const Basic: Story = {
     args: {},
+};
+
+export const WithEntranceAnimation: Story = {
+    args: {},
+    render: (args) => {
+        return {
+            props: {
+                ...args,
+                show: true,
+            },
+            template: `
+            <div class="flex flex-col min-h-96 min-w-96 gap-4">
+                <button (click)="show = !show" class="p-6 bg-light-gray rounded-full cursor-pointer hover:shadow-inner transition-all border border-gray">toggle content</button>
+                <p>Is empty: {{isEmpty}}</p>
+                <div puibeHideIfEmptyText (emptyTextChange)="isEmpty = $event">
+
+                    <component-with-entrace-animation [show]="show"/>
+
+                </div>
+            </div>`,
+        };
+    },
+};
+
+export const WithEntranceAnimationAndNestedDirectives: Story = {
+    args: {},
+    render: (args) => {
+        return {
+            props: {
+                ...args,
+                show: true,
+            },
+            template: `
+            <div class="flex flex-col min-h-96 min-w-96 gap-4">
+                <button (click)="show = !show" class="p-6 bg-light-gray rounded-full cursor-pointer hover:shadow-inner transition-all border border-gray">toggle content</button>
+                <p>Is child empty: {{isEmpty}}, Is parent empty: {{isParentEmpty}}</p>
+                <div puibeHideIfEmptyText (emptyTextChange)="isParentEmpty = $event">
+                    <div puibeHideIfEmptyText (emptyTextChange)="isEmpty = $event">
+
+                        <component-with-entrace-animation [show]="show"/>
+
+                    </div>
+                </div>
+            </div>`,
+        };
+    },
 };